+++ nilearn/decomposition/base.py
@@ -107,7 +107,7 @@ def mask_and_reduce(masker, imgs,
         See http://nilearn.github.io/manipulating_images/input_output.html
         List of subject data to mask, reduce and stack.
 
-    confounds: CSV file path or 2D matrix, optional
+    confounds: CSV file path or numpy ndarray, or pandas DataFrame, optional
         This parameter is passed to signal.clean. Please see the
         corresponding documentation for details.
 
@@ -359,7 +359,8 @@ class BaseDecomposition(BaseEstimator, CacheMixin, TransformerMixin):
             Data on which the mask is calculated. If this is a list,
             the affine is considered the same for all.
 
-        confounds : list of CSV file paths or 2D matrices
+        confounds : list of CSV file paths or numpy.ndarrys
+                    or pandas DataFrames, optional,
             This parameter is passed to nilearn.signal.clean. Please see the
             related documentation for details. Should match with the list
             of imgs given.
@@ -428,7 +429,8 @@ class BaseDecomposition(BaseEstimator, CacheMixin, TransformerMixin):
             See http://nilearn.github.io/manipulating_images/input_output.html
             Data to be projected
 
-        confounds: CSV file path or 2D matrix
+        confounds: CSV file path or numpy.ndarray or pandas DataFrame,
+                   optional,
             This parameter is passed to nilearn.signal.clean. Please see the
             related documentation for details
 
@@ -505,7 +507,8 @@ class BaseDecomposition(BaseEstimator, CacheMixin, TransformerMixin):
             See http://nilearn.github.io/manipulating_images/input_output.html
             Data to be scored
 
-        confounds: CSV file path or 2D matrix
+        confounds: CSV file path or numpy.ndarray or pandas DataFrame,
+                   optional,
             This parameter is passed to nilearn.signal.clean. Please see the
             related documentation for details