+++ nilearn/glm/tests/test_first_level.py
@@ -274,6 +274,13 @@ def test_fmri_inputs():
                 FirstLevelModel(mask_img=None).fit([fi], design_matrices=d)
                 FirstLevelModel(mask_img=mask).fit(fi, design_matrices=[d])
                 FirstLevelModel(mask_img=mask).fit([fi], design_matrices=[d])
+                # test with confounds
+                FirstLevelModel(mask_img=mask).fit([fi], design_matrices=[d],
+                                                   confounds=conf)
+                # test with confounds as numpy array
+                FirstLevelModel(mask_img=mask).fit([fi], design_matrices=[d],
+                                                   confounds=conf.values)
+                
                 FirstLevelModel(mask_img=mask).fit([fi, fi],
                                                    design_matrices=[d, d])
                 FirstLevelModel(mask_img=None).fit((fi, fi),